Create module to aid in shared ca creation
This commit adds the initial code and README for the new shared_ca
class. This class simply automates some of the PE 2.5 pieces that
one must tweak in order to create a shared CA environment, where
the console host acts as the CA and each master gets a copy of that
CA so that they can communicate back to the console but sign their
own agents.

This module is meant to be used with puppet apply on a machine that
has a fresh install of PE 2.5 and is intended to only be used to aid
in getting the machine back into a working state with the shared CA.

See the README for details.