Improve ordering and docs.
This commit vastly improves the README-based documentation as to
this modules purpose, what it achieves, how it does it and what
the user is expected to do.

Based on experienced bugs with resource ordering, explicit
dependencies were added between some resources and an exec blowing
out the existing CA directory prior to replacing it with the shared
console CA was added cleanliness.

To be clear, the module at this point has received about an hours
worth of work to get a customer up and running. Improvements are
certainly needed for functionality, automation and 'doing it the
right way'.